services/data-ingestion/requirements.txt
kafka-python==2.0.2
redis==5.0.1
websocket-client==1.7.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
yfinance==0.2.33
polygon-api-client==1.12.5
alpha-vantage==2.3.1
kiteconnect==4.2.0
upstox-python-sdk==2.7.0
finnhub-python==2.4.18
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
prometheus-client==0.19.0
